move LogoChain to separate dir
