+++ client/my-sites/importer/newsletter/importer.tsx
@@ -16,7 +16,7 @@ import { useAnalyzeUrlQuery } from 'calypso/data/site-profiler/use-analyze-url-q
 import { useSelector } from 'calypso/state';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
 import Content from './content';
-import { LogoChain } from './logo-chain';
+import { LogoChain } from './logo-chain/logo-chain';
 import SelectNewsletterForm from './select-newsletter-form';
 import Subscribers from './subscribers';
 import Summary from './summary';
